feat(test_actions.py): streamline data handling in test_simulation
This commit simplifies the data handling in the test_simulation
function.Instead of storing all run data in a DataFrame and writing it
to a fileat the end, each run's data is now immediately written to the
file. Thischange reduces memory usage and the risk of data loss in case
of an error.The 'TODO Fix me' comment was also removed as it is no
longer relevant.